feat: implement getAllUserAuctions method and remove useGetAddressAuctionHistory hook
